Add line clear function - 알 수 없는 렉 현상이 있음. 아마 게임 로직에 문제가 있는듯..

+++ Linux/src/chronometry.h
@@ -8,7 +8,7 @@
 #include <unistd.h>
 #include <assert.h>
 
-typedef double __sseconds_t;
+typedef float __sseconds_t;
 typedef __syscall_slong_t __snseconds_t;
 
 #define NSEC_VAL ((__snseconds_t)1e9)
